refactor(webhook): split apply_handlers.go by lifecycle stage
Move the 975-line pkg/webhook/apply_handlers.go into four focused files
within the same package. Pure file-organization refactor — no logic,
signature, or visibility changes.

- apply_handlers.go: command entry points (handleApplyCommand,
  handleApplyConfirmCommand, handleUnlockCommand)
- apply_execute.go: shared executeApply core + ddlMatchesStoredPlan
  drift detection
- apply_gating.go: PR-status gating (enforcePassingChecks +
  filterFailing/InProgress/required-rollup helpers)
- apply_check_records.go: apply-lifecycle check record updates
  (storeApplyPlanCheckRecord, updateCheckRecordForApplyStart,
  updateCheckRunAfterUnlock)

Validation:
- go build ./...
- go test -count=1 -race ./pkg/webhook/...
- make test-unit (all packages pass)
- golangci-lint run --new-from-rev=origin/main ./pkg/webhook/... (0 issues)
